Properly document and place the new code
