+++ crates/terminal/src/terminal.rs
@@ -2,7 +2,6 @@ pub mod mappings;
 
 pub use alacritty_terminal;
 
-mod grid;
 mod pty_info;
 pub mod terminal_settings;
 
@@ -29,7 +28,6 @@ use futures::{
     FutureExt,
 };
 
-use grid::display_task_results;
 use mappings::mouse::{
     alt_scroll, grid_point, grid_point_and_side, mouse_button_report, mouse_moved_report,
     scroll_report,
@@ -1446,7 +1444,64 @@ impl Terminal {
                 task.status.register_terminal_exit();
             }
         };
-        display_task_results(self.term.as_ref(), task, error_code);
+        let escaped_full_label = task.full_label.replace("\r\n", "\r").replace('\n', "\r");
+        let task_line = match error_code {
+            Some(0) => {
+                format!("{TASK_DELIMITER}Task `{escaped_full_label}` finished successfully")
+            }
+            Some(error_code) => {
+                format!("{TASK_DELIMITER}Task `{escaped_full_label}` finished with non-zero error code: {error_code}")
+            }
+            None => {
+                format!("{TASK_DELIMITER}Task `{escaped_full_label}` finished")
+            }
+        };
+        let escaped_command_label = task.command_label.replace("\r\n", "\r").replace('\n', "\r");
+        let command_line = format!("{TASK_DELIMITER}Command: '{escaped_command_label}'");
+        // SAFETY: the invocation happens on non `TaskStatus::Running` tasks, once,
+        // after either `AlacTermEvent::Exit` or `AlacTermEvent::ChildExit` events that are spawned
+        // when Zed task finishes and no more output is made.
+        // After the task summary is output once, no more text is appended to the terminal.
+        unsafe { append_text_to_term(&mut self.term.lock(), &[&task_line, &command_line]) };
+    }
+}
+
+const TASK_DELIMITER: &str = "⏵ ";
+
+/// Appends a stringified task summary to the terminal, after its output.
+///
+/// SAFETY: This function should only be called after terminal's PTY is no longer alive.
+/// New text being added to the terminal here, uses "less public" APIs,
+/// which are not maintaining the entire terminal state intact.
+///
+///
+/// The library
+///
+/// * does not increment inner grid cursor's _lines_ on `input` calls
+/// (but displaying the lines correctly and incrementing cursor's columns)
+///
+/// * ignores `\n` and \r` character input, requiring the `newline` call instead
+///
+/// * does not alter grid state after `newline` call
+/// so its `bottommost_line` is always the the same additions, and
+/// the cursor's `point` is not updated to the new line and column values
+///
+/// * ??? there could be more consequences, and any further "proper" streaming from the PTY might bug and/or panic.
+/// Still, concequent `append_text_to_term` invocations are possible and display the contents correctly.
+///
+/// Despite the quirks, this is the simplest approach to appending text to the terminal: its alternative, `grid_mut` manipulations,
+/// do not properly set the scrolling state and display odd text after appending; also those manipulations are more tedious and error-prone.
+/// The function achieves proper display and scrolling capabilities, at a cost of grid state not properly synchronized.
+/// This is enough for printing moderately-sized texts like task summaries, but might break or perform poorly for larger texts.
+unsafe fn append_text_to_term(term: &mut Term<ZedListener>, text_lines: &[&str]) {
+    term.newline();
+    term.grid_mut().cursor.point.column = Column(0);
+    for line in text_lines {
+        for c in line.chars() {
+            term.input(c);
+        }
+        term.newline();
+        term.grid_mut().cursor.point.column = Column(0);
     }
 }
